requirements.txt
# Requirements for Dockerfile
ipykernel
apache-airflow-providers-papermill
